Book database
This directory contains 180 books of different genres. There are six genres:
-Mathematics
-Law
-Biology
-Journalism
-Psychology
-General literature

There are 30 books of each genre, all in .txt format for later processing. For the Naive-Bayes classifier, 20 books of each genre
should be used for the training process. 10 books of each genre should be used for the validation process.